add documentation and structure to level tutorials Signed-off-by: dbulahov <bulahovdenis@gmail.com>
