+++ instance/emf/serializer/pom.xml
@@ -526,17 +526,23 @@
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
       <artifactId>
-        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchy.levelifblankname</artifactId>
+        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.level.ifblankname</artifactId>
       <version>${project.version}</version>
     </dependency>
 
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
       <artifactId>
-        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchy.levelifparentsname</artifactId>
+        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.level.ifblanknamemultiple</artifactId>
       <version>${project.version}</version>
     </dependency>
 
+    <dependency>
+      <groupId>org.eclipse.daanse</groupId>
+      <artifactId>
+        org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.level.ifparentsname</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
     <dependency>
       <groupId>org.eclipse.daanse</groupId>
