--- instance/emf/tutorial/hierarchy/levelifblankname/src/main/resources/data/Level_2_Multiple.csv
@@ -1,9 +1,0 @@
-"KEY","NAME","L1_KEY"
-INTEGER,VARCHAR,INTEGER
-1,AA,1
-2,AB,1
-3,,2
-4,,3
-5,DA,4
-6,,4
-7,,4
