--- instance/emf/tutorial/hierarchy/levelifblankname/src/main/resources/data/Level_3_Multiple.csv
@@ -1,11 +1,0 @@
-"KEY","NAME","L2_KEY"
-INTEGER,VARCHAR,INTEGER
-1,AAA,1
-2,AAB,1
-3,,2
-4,,3
-5,CAA,4
-6,CAB,4
-7,DAA,5
-8,,5
-9,,6
