+++ instance/emf/tutorial/level/ifblankname/pom.xml
@@ -10,21 +10,12 @@
 * SPDX-License-Identifier: EPL-2.0
 **********************************************************************/
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
     <groupId>org.eclipse.daanse</groupId>
-    <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial</artifactId>
+    <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.level</artifactId>
     <version>${revision}</version>
   </parent>
-  <packaging>pom</packaging>
-
-  <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.hierarchy</artifactId>
-
-  <modules>
-    <module>inlinetable</module>
-    <module>view</module>
-  </modules>
+  <artifactId>org.eclipse.daanse.rolap.mapping.instance.emf.tutorial.level.ifblankname</artifactId>
 </project>
