+++ instance/emf/tutorial/level/ifblankname/src/main/resources/data/Level_2_NULL.csv
@@ -1,0 +1,5 @@
+"KEY","NAME","L1_KEY"
+INTEGER,VARCHAR,INTEGER
+1,AA,1
+2,AB,1
+3,NULL,2
