+++ instance/emf/tutorial/level/ifblanknamemultiple/src/main/resources/data/Level_2_Multiple.csv
@@ -1,0 +1,9 @@
+"KEY","NAME","L1_KEY"
+INTEGER,VARCHAR,INTEGER
+1,AA,1
+2,AB,1
+3,NULL,2
+4,NULL,3
+5,DA,4
+6,NULL,4
+7,NULL,4
