+++ instance/emf/tutorial/level/ifblanknamemultiple/src/main/resources/data/Level_3_Multiple.csv
@@ -1,0 +1,11 @@
+"KEY","NAME","L2_KEY"
+INTEGER,VARCHAR,INTEGER
+1,AAA,1
+2,AAB,1
+3,NULL,2
+4,NULL,3
+5,CAA,4
+6,CAB,4
+7,DAA,5
+8,NULL,5
+9,NULL,6
